feat(sui-domain): handle all feedback
