--- packages/sui-domain/fetcher/factory.js
@@ -1,6 +1,0 @@
-import AxiosFetcher from './AxiosFetcher'
-
-export default class FetcherFactory {
-  static axiosFetcher = ({config}) =>
-    new AxiosFetcher({config})
-}
