+++ packages/sui-domain/package.json
@@ -2,14 +2,16 @@
   "name": "@s-ui/domain",
   "version": "0.0.1-beta.0",
   "description": "SDK for creating domains",
+  "main": "/lib",
   "scripts": {
-    "lib": "rm -Rf ./lib && mkdir -p ./lib && babel --presets sui ./src --out-dir ./lib",
-    "prepublish": "npm run lib && node scripts/prepublish",
-    "postpublish": "node scripts/postpublish"
+    "lib": "rm -Rf ./lib && mkdir -p ./lib && babel --presets sui ./src --out-dir ./lib"
   },
   "keywords": [],
-  "author": "",
+  "author": "Enablers Frontend",
   "license": "ISC",
+  "dependencies": {
+    "axios": "0.17.1"
+  },
   "devDependencies": {
     "babel-cli": "6.24.1",
     "babel-preset-sui": "1",
