--- packages/sui-domain/src/Fetcher.js
@@ -1,1 +1,0 @@
-export * as FetcherFactory from './fetcher/factory'
