+++ packages/sui-domain/src/fetcher/factory.js
@@ -1,6 +1,6 @@
 import AxiosFetcher from './AxiosFetcher'
 
 export default class FetcherFactory {
-  static axiosFetcher = ({config}) =>
+  static httpFetcher = ({config}) =>
     new AxiosFetcher({config})
 }
