--- packages/sui-domain/src/helpers/required.js
@@ -1,3 +1,0 @@
-export default (param) => {
-  throw new Error(`${param} is required`)
-}
