+++ packages/sui-domain/src/index.js
@@ -1,24 +1,8 @@
-
-import required from './helpers/required'
-
-export default class EntryPoint {
-  constructor ({ config = required('config'), useCases = require('useCases') }) {
-    this._config = config
-    this._useCases = useCases
-  }
-
-  get (key) {
-    if (key === 'config') {
-      return this._config
-    }
-
-    return this._useCases[key] || {
-      execute: () => Promise.reject(new Error(`[EntryPoint#get] ${key} not defined`))
-    }
-  }
-
-  config (key, value) {
-    this._config.set(key, value)
-    return this
-  }
-}
+export Entity from './Entity'
+export EntryPoint from './EntryPoint'
+export FetcherFactory from './fetcher/factory'
+export Mapper from './Mapper'
+export Repository from './Repository'
+export Service from './Service'
+export UseCase from './UseCase'
+export ValueObject from './ValueObject'
